HBX-1681: Move class 'org.hibernate.tool.hbm2x.visitor.HBMTagForValueVisitor' to 'org.hibernate.tool.internal.export.hbm.HBMTagForValueVisitor'
