+++ orm/src/main/java/org/hibernate/tool/internal/export/hbm/Cfg2HbmTool.java
@@ -44,7 +44,6 @@ import org.hibernate.persister.entity.UnionSubclassEntityPersister;
 import org.hibernate.tool.hbm2x.ExporterException;
 import org.hibernate.tool.hbm2x.visitor.EntityNameFromValueVisitor;
 import org.hibernate.tool.hbm2x.visitor.HBMTagForPersistentClassVisitor;
-import org.hibernate.tool.hbm2x.visitor.HBMTagForValueVisitor;
 import org.hibernate.tool.internal.util.SkipBackRefPropertyIterator;
 
 /**
